Fixed bug related to using the wrong q in steiner classical dreyfus related to relabeling, basically the inverse labeling was required

+++ common.h
@@ -8,5 +8,5 @@
 #include "boost/multi_array.hpp"
 typedef boost::multi_array<int, 2> adjancy_matrix;
 typedef adjancy_matrix::index index;
-typedef  uint_fast32_t set_t;
+typedef  uint_fast64_t set_t;
 #endif //FAST_SUBSET_CONVOLUTION_COMMON_H
